+++ src/SWP/Bundle/SettingsBundle/Exception/InvalidScopeException.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Superdesk Web Publisher Settings Bundle.
+ *
+ * Copyright 2017 Sourcefabric z.ú. and contributors.
+ *
+ * For the full copyright and license information, please see the
+ * AUTHORS and LICENSE files distributed with this source code.
+ *
+ * @copyright 2017 Sourcefabric z.ú
+ * @license http://www.superdesk.org/license
+ */
+
+namespace SWP\Bundle\SettingsBundle\Exception;
+
+class InvalidScopeException extends \InvalidArgumentException
+{
+    public function __construct($scope)
+    {
+        parent::__construct(sprintf('Scope %s is not supported!.', $scope));
+    }
+}