+++ src/SWP/Bundle/SettingsBundle/Tests/Functional/EventListener/ResourceResponseListener.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Superdesk Web Publisher Settings Bundle.
+ *
+ * Copyright 2017 Sourcefabric z.ú. and contributors.
+ *
+ * For the full copyright and license information, please see the
+ * AUTHORS and LICENSE files distributed with this source code.
+ *
+ * @copyright 2017 Sourcefabric z.ú
+ * @license http://www.superdesk.org/license
+ */
+
+namespace SWP\Bundle\SettingsBundle\Tests\Functional\EventListener;
+
+use SWP\Component\Common\Response\ResourcesListResponseInterface;
+use SWP\Component\Common\Response\SingleResourceResponseInterface;
+use Symfony\Component\Form\FormInterface;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Event\GetResponseForControllerResultEvent;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\NameConverter\CamelCaseToSnakeCaseNameConverter;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Serializer\Serializer;
+use Symfony\Component\Serializer\SerializerInterface;
+
+final class ResourceResponseListener
+{
+    /**
+     * @param GetResponseForControllerResultEvent $event
+     */
+    public function onKernelView(GetResponseForControllerResultEvent $event)
+    {
+        $controllerResult = $event->getControllerResult();
+        if ($controllerResult instanceof ResourcesListResponseInterface) {
+            $event->setResponse(new JsonResponse($controllerResult->getResources()));
+        } elseif ($controllerResult instanceof SingleResourceResponseInterface) {
+            if ($controllerResult->getResource() instanceof FormInterface) {
+                $errors = array();
+                foreach ($controllerResult->getResource()->getErrors(true) as $error) {
+                    $errors[] = array('message' => $error->getMessage());
+                }
+
+                $event->setResponse(new Response($this->getSerializer()->serialize($errors, 'json')));
+
+                return;
+            }
+
+            $event->setResponse(new Response($this->getSerializer()->serialize($controllerResult->getResource(), 'json')));
+        }
+    }
+
+    /**
+     * @return SerializerInterface
+     */
+    private function getSerializer(): SerializerInterface
+    {
+        $encoders = [new JsonEncoder()];
+        $normalizer = new ObjectNormalizer(null, new CamelCaseToSnakeCaseNameConverter());
+        $normalizers = [$normalizer];
+
+        return new Serializer($normalizers, $encoders);
+    }
+}