+++ CHANGELOG.md
@@ -8,6 +8,9 @@ in 0.1 minor versions.
 To get the diff for a specific change, go to https://github.com/superdesk/web-publisher/commit/XXX where XXX is the change hash
 
 * 0.6.0
+ * feature [#312] Improve user registration 
+ * feature [#312] Add Settings Bundle
+ * feature [#309] Validate objects based on configured model's validation when content is pushed
  * feature [#295] Implemented article preview for users with special privileges
  * feature [#297] Add API to delete article
  * feature [#294] Add support for Manual Content List in API